third_party/blink/web_tests/external/wpt/html/cross-origin-opener-policy/popup-coop-by-sw-from-coop.https.html

<html>
<head>
  <title>
    Check ServiceWorkers can set COOP/COEP headers for documents's responses
  </title>
  <meta name="timeout" content="long">
  <meta name="variant" content="?1-1">
  <meta name="variant" content="?2-2">
  <meta name="variant" content="?3-3">
  <meta name="variant" content="?4-last">
  <script src="/resources/testharness.js"></script>
  <script src="/resources/testharnessreport.js"></script>
  <script src="/common/utils.js"></script>
  <script src="/common/subset-tests.js"></script>
  <script src="/service-workers/service-worker/resources/test-helpers.sub.js">
  </script>
  <script src="./resources/popup-coop-by-sw.js"> </script>
</head>
<body>
<script>

// The current document have COOP/COEP defined. A popup will be opened.
// Its response will be provided by a ServiceWorker. A browsing context
// group swap must happen if the popup's headers contains COOP/COEP headers.

subsetTest(popupCoopBySwTest, "sw_normal, popup_normal",
                               sw_normal, popup_normal,
                               swap_browsing_context_group)

subsetTest(popupCoopBySwTest, "sw_normal, popup_coop",
                               sw_normal, popup_coop,
                               keep_browsing_context_group)

subsetTest(popupCoopBySwTest, "sw_coop, popup_normal",
                               sw_coop, popup_normal,
                               swap_browsing_context_group)

subsetTest(popupCoopBySwTest, "sw_coop, popup_coop",
                               sw_coop, popup_coop,
                               keep_browsing_context_group)

</script>
</body>
</html>
