LayoutTests/imported/w3c/web-platform-tests/html/cross-origin-opener-policy/popup-coop-by-sw-from-coop.https-expected.txt


FAIL sw_normal, popup_normal assert_equals: expected "No opener is set" but got "An opener is set"
PASS sw_normal, popup_coop
FAIL sw_coop, popup_normal assert_equals: expected "No opener is set" but got "An opener is set"
PASS sw_coop, popup_coop
